requirements.txt
Flask
openpyxl
supabase
python-dotenv
flask-login
